pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
]
